+++ dongibuyeo/src/main/java/com/shinhan/dongibuyeo/domain/member/controller/MemberController.java
@@ -1,0 +1,17 @@
+package com.shinhan.dongibuyeo.domain.member.controller;
+
+import com.shinhan.dongibuyeo.domain.member.service.MemberService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/member")
+public class MemberController {
+
+    private final MemberService memberService;
+
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+}
