+++ dongibuyeo/src/main/java/com/shinhan/dongibuyeo/domain/member/dto/request/MemberRequest.java
@@ -1,0 +1,4 @@
+package com.shinhan.dongibuyeo.domain.member.dto.request;
+
+public class MemberRequest {
+}
