+++ dongibuyeo/src/main/java/com/shinhan/dongibuyeo/domain/member/dto/response/MemberResponse.java
@@ -1,0 +1,4 @@
+package com.shinhan.dongibuyeo.domain.member.dto.response;
+
+public class MemberResponse {
+}
