+++ dongibuyeo/src/main/java/com/shinhan/dongibuyeo/domain/member/mapper/MemberMapper.java
@@ -1,0 +1,5 @@
+package com.shinhan.dongibuyeo.domain.member.mapper;
+
+@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
+public interface MemberMapper {
+}
