+++ dongibuyeo/src/main/java/com/shinhan/dongibuyeo/domain/member/repository/MemberRepository.java
@@ -1,9 +1,11 @@
-package com.shinhan.dongibuyeo.member.repository;
+package com.shinhan.dongibuyeo.domain.member.repository;
 
-import com.shinhan.dongibuyeo.member.entity.Member;
+import com.shinhan.dongibuyeo.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
 
 public interface MemberRepository extends JpaRepository<Member, UUID> {
+
+
 }
