--- dongibuyeo/src/main/java/com/shinhan/dongibuyeo/member/controller/MemberController.java
@@ -1,17 +1,0 @@
-package com.shinhan.dongibuyeo.member.controller;
-
-import com.shinhan.dongibuyeo.member.service.MemberSevice;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/member")
-public class MemberController {
-
-    private final MemberSevice memberSevice;
-
-    public MemberController(MemberSevice memberSevice) {
-        this.memberSevice = memberSevice;
-    }
-
-}
