--- dongibuyeo/src/main/java/com/shinhan/dongibuyeo/member/dto/request/MemberRequest.java
@@ -1,4 +1,0 @@
-package com.shinhan.dongibuyeo.member.dto.request;
-
-public class MemberRequest {
-}
