--- dongibuyeo/src/main/java/com/shinhan/dongibuyeo/member/dto/response/MemberResponse.java
@@ -1,4 +1,0 @@
-package com.shinhan.dongibuyeo.member.dto.response;
-
-public class MemberResponse {
-}
